Optimize README load time.
I have inlined the profile badges in order to optimize loading time.
A new `Makefile` is added. Running `make` will update the images.
Users visiting your profile will experience a better page load time, as
the content is served from GitHub directly, and thus does not need to
open more HTTPS connections.

The `make` command should be run in a cronjob with daily updates pushed
to this repository. This is out of scope for this commit.